requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
motor>=3.2.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
